refactor!: move Scope to kubernetes-model-common and clean it up
